Init code rearrangement
